Created a general tab flow structure for Session recording, a Session Retrospective/calendar/dashboard, and a Device management tab.
