+++ mobile/ios-app/Sources/UI/MainView.swift
@@ -2,8 +2,21 @@ import SwiftUI
 
 struct MainView: View {
     var body: some View {
-        NavigationView {
-            ConnectView()
+        TabView {
+            SessionView()
+                .tabItem {
+                    Label("Session", systemImage: "record.circle")
+                }
+            
+            CalendarView()
+                .tabItem {
+                    Label("Dashboard", systemImage: "calendar")
+                }
+            
+            PairDeviceView()
+                .tabItem {
+                    Label("Device", systemImage: "antenna.radiowaves.left.and.right")
+                }
         }
     }
 }
