+++ mobile/ios-app/Sources/UI/Session/LiveSession/PulseRectangle.swift
